Fix economy, prices, policy export and turn-processing speed
Economy:
- Net income was credited to the treasury twice per turn (processEconomy
  and recordIncomeSnapshot both applied it), so every country earned
  double the income shown in the UI.
- Treasury is now double. As a 32-bit float it could not represent
  consecutive integers above ~16.7M, so 'treasury += net' silently
  became a no-op once a country got rich and the economy froze.

Prices:
- The Specialize button derived its price from the upgrade cost, which
  at max industry level was a 99999 'no next level' sentinel, showing
  ~$150,000 and making specialisation impossible. It is now priced off
  the current level.
- Industry cost research is registered as a positive percentage, but the
  modifier added it, so researching 'Industry cost -50%' raised prices
  to 1.5x. Now subtracts.
- Recruit cost queried 'armyCostPct', which is not a real effect name and
  always returned 0; the real one is conscriptionCostPct.
- Industry level tables are now clamped at both ends; level comes
  unvalidated from map data and could index them out of bounds.

Turn processing:
- The embedded 13MB map.odmap was re-deflated at max level on every save
  rewrite, despite already being compressed. Now stored.
- The save archive was fully rebuilt twice per turn; appendTurn can now
  write the state snapshot in the same pass.
- Artillery orders were processed twice per turn, applying all damage
  twice.

Map editor:
- starting_minority_policies.json was generated but never added to the
  exported .odmap, silently discarding all Ethnic Relations settings.
- Ethnic settings for hand-added countries were keyed by an empty ISO
  and dropped at export; ISO codes are now assigned up front.

+++ src/Game_UI.cpp
@@ -442,8 +442,8 @@ void Game::drawCeasefireScreen() {
         curY += 16;
     };
 
-    int pMax = (int)std::max(0.0f, m_countries.getAll()[m_playerCountryId].treasury + 1000);
-    int tMax = (int)std::max(0.0f, targetC->treasury + 1000);
+    int pMax = (int)std::max(0.0, m_countries.getAll()[m_playerCountryId].treasury + 1000);
+    int tMax = (int)std::max(0.0, targetC->treasury + 1000);
     drawMoneySlider("Money we offer", m_ceasefireOurMoney, pMax, Color{40, 200, 40, 220});
     drawMoneySlider("Money we demand", m_ceasefireTheirMoney, tMax, Color{220, 60, 60, 220});
 
@@ -642,7 +642,7 @@ void Game::updateCeasefireScreen() {
 
     if (CheckCollisionPointRec(mouse, sendBtn) && click) {
         // Clamp offer to what the player can actually pay (treasury >= 0)
-        float& pTreas = m_countries.getAll()[m_playerCountryId].treasury;
+        double& pTreas = m_countries.getAll()[m_playerCountryId].treasury;
         if (m_ceasefireOurMoney > (int)pTreas) m_ceasefireOurMoney = (int)pTreas;
         if (m_ceasefireOurMoney < 0) m_ceasefireOurMoney = 0;
         // Deduct offered money from treasury immediately (refunded if rejected)
